updated logic for dependent tables

+++ src/controllers/venue.js
@@ -1,4 +1,5 @@
 const venueSchema = require('../models/venue');
+const eventSchema = require('../models/events');
 const errors = require('../config/errors.json');
 
 const fetchVenue = async (req, res, next) => {
@@ -45,6 +46,13 @@ const updateVenue = async (req, res, next) => {
 
 const deleteVenue = async (req, res, next) => {
     try {
+        const eventDetails = await eventSchema.find({ venueId: req.params.id });
+        if (eventDetails){
+            throw {
+                ...errors[404],
+                data: `cannot delete venue: ${req.params.id}. Venue is already linked to an event`
+            } 
+        }
         const deletedVenueDetails = await venueSchema.deleteOne({ _id: req.params.id });
         if (!deletedVenueDetails.deletedCount) {
             throw {
